+++ derive_builder/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.12.0] - 2022-11-28
 - Produce error when `default` is used with `field(type = "...")` rather than silently ignoring `default` #269
 - Add support for `crate = "..."` to support re-export scenarios #274